+++ demo/src/main/java/com/tuture/demo/model/dto/SignInDto.java
@@ -1,0 +1,31 @@
+package com.tuture.demo.model.dto;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Pattern;
+import lombok.*;
+
+public class SignInDto {
+
+    @Getter
+    @Builder
+    @AllArgsConstructor
+    @NoArgsConstructor
+    public static class Request {
+        @NotBlank(message = "이메일은 필수 입력 값입니다.")
+        @Pattern(regexp = "^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,6}$", message = "이메일 형식이 올바르지 않습니다.")
+        private String email;
+
+        @NotBlank(message = "비밀번호는 필수 입력 값입니다.")
+        private String password;
+    }
+
+    @Getter
+    @Builder
+    public static class Response {
+        private long userId;
+        private String email;
+        private String nickname;
+        private String access_token;
+    }
+
+}
