+++ demo/src/main/java/com/tuture/demo/service/BoardServiceImpl.java
@@ -1,7 +1,7 @@
 package com.tuture.demo.service;
 
 import com.tuture.demo.model.dao.BoardDao;
-import com.tuture.demo.model.dto.Board;
+import com.tuture.demo.model.domain.Board;
 import com.tuture.demo.model.dto.AddBoardDto;
 import com.tuture.demo.model.dto.SearchCondition;
 import lombok.RequiredArgsConstructor;
@@ -18,8 +18,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class BoardServiceImpl implements BoardService{
 
-    @Autowired
-    private BoardDao boardDao;
+    private final BoardDao boardDao;
 
     @Override
     public List<Board> getBoardList(){
